developer guide: restructure howtos, tech notes and dev guides.
This is a large cleanup that mostly moves files into the correct
directories based on subject.

It moves howtos/process to contribute/process, creates a developer
guides firmware directory and moves all FW development topics to this dierectory.

Finally it adds some place holders for new files that are WiP.

Signed-off-by: Liam Girdwood <liam.r.girdwood@linux.intel.com>