Makefile: add ../sof/doc/doxygen/ dependency
Doxygen is not incremental but it is super fast in this project with
this configuration, so run it every time before we run sphinx. Totally
negligible compared to the minutes required by plantUML.jar.

Keep it optional not to add new dependencies. Now for users who opt-in,
this commit will keep showing the doxygen warnings on every incremental
build (unlike the .rst warnings unfortunately).

Print the magic cmake invocation so no one has to go and read the build
HOWTO to get (re)started.

Signed-off-by: Marc Herbert <marc.herbert@intel.com>