build-guide: Use env var for referring ~/work/sof
Currently, if one wishes to follow the documentation but has a different
directory organization for their code they have to replace all the
occurences of ~/work/sof in the documentation with their directory. By
using $SOF_WORKSPACE the user can set the root once and copy paste all
the commands from the documentation as is.

Signed-off-by: Shreeya Patel <shreeya.patel23498@gmail.com>