contributing: Add section on process for SOF FW ABI changes
Document revised process for ABI changes including a new step
for Request-for-Comments phase. This is intended to ensure
early engagement of interface changes across driver and firmware
teams, to avoid wasting effort when concerns are raised at
the very end of the process.

Add cross-links between kernel driver and firmware sections.
Also add link to ABI versioning definition in kernel/abi.h
and pointers to TSC and driver maintainer team and who are
the members.

Signed-off-by: Kai Vehmanen <kai.vehmanen@linux.intel.com>