release: update release information
Remove obsolete reference to v1.3 as the latest release. It's better
not to mention the latest here at all and instead just point readers
to the GitHub release page.

Also remove text referring to linux-firmware.git and instead
refer to sof-bin repository as the official channel for binary
release packages.

Signed-off-by: Kai Vehmanen <kai.vehmanen@linux.intel.com>