Makefile: new SOF_DOC_BUILD un-hardcodes ../sof/doc
The default and recommended location for the doxygen build is a source
directory, which gets polluted by the build and incredibly tedious
to (dist)clean manually. Don't change that default and documentation
yet (baby steps) but make it at least possible to build (and clean)
doxygen outside the source directory per the usual CMake practices.

As a side-effect this makes the connection with doxygen and breathe less
implicit in the Makefile.

Signed-off-by: Marc Herbert <marc.herbert@intel.com>