feat(getting_started): update organization and instructions
- Change the organization of the getting started section to make it
easier to determine which guide to follow.
- Add instructions for building a custom kernel with SOF config so that
it can be booted on that same device. This is useful for people
without dedicated test hardware who just want to try out the latest
code, such as when they are running into issues with new hardware.
- Revamp the ktest setup documentation. This was very out of date and
had tons of errors, especially for using Fedora. It has now been validated
to work on both Fedora and Ubuntu target devices, from a Ubuntu dev
machine (but Fedora should work just fine too).

Signed-off-by: Noah Klayman <noah.klayman@intel.com>