arch: fw: cleanup links in memory management overview
Initial description was unbalanced. One section contained links
while the other had the links moved to the Read More section below
the diagram.

Now all the links are moved to the Read More. External links are
annotated to let the reader know that opening results in move
to another website.

Signed-off-by: Marcin Maka <marcin.maka@intel.com>