intel_debug: introduction: Update default IPC4 topology path with example
We need to keep the topology files for different ADSP generations separate
because they can include NHLT blobs and the blobs are not compatible among
different versions.

While this only affects generically named topology files
(sof-hda-generic*.tplg), it is cleaner to expect the topology files to be
stored under different directories rather than having some in sof-ipc4-tplg
root and some under directories.
Keeping the generational topology files in separate directory also enables
us to offer backwards compatibility for deployed MTL installs with older
kernels.

The new default topology path shall be:
/lib/firmware/intel/sof-ipc4-tplg/PLAT/sof-CONFIG.tplg

With a practical note that the PLAT is symlinked to a common directory for
the generation they belopng to:
.
└── intel
    ├── sof-ace-tplg -> sof-ipc4-tplg/ACE1
    └── sof-ipc4-tplg
        ├── ACE1
        ├── CAVS2.5
        ├── mtl -> ACE1
        ├── tgl -> CAVS2.5
        └── tgl-h -> CAVS2.5

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>