intel_debug: introduction: Detail description of a modular firmware release
Separate the firmware lookup table for non-modular and modular SOF releases
to be able to document the locations and file names the firmware will
be looking for the individual files.

Extend the description of the two type of SOF release and convert the
list-table to a normal table for better descriptions for the
configurations.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>